% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{generateCohortLookup}
\alias{generateCohortLookup}
\title{Generate cohort lookup table from cohortDefinitionSet/cohortCounts}
\usage{
generateCohortLookup(cohortDef)
}
\arguments{
\item{cohortDef}{Data frame containing at least cohortId and cohortName (cohortDefinitionSet)}
}
\value{
data.frame with cohortId, cohortName, cancer_type, subset, category
}
\description{
Generate cohort lookup table from cohortDefinitionSet/cohortCounts
}
