% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCohortDiagnostics.R
\name{runCohortDiagnostics}
\alias{runCohortDiagnostics}
\title{Run Cohort Diagnostics}
\usage{
runCohortDiagnostics(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortDefinitionSet,
  exportFolder
)
}
\description{
Executes CohortDiagnostics for all generated breast cancer cohorts.
}
