% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCohortGeneration.R
\name{runCohortGeneration}
\alias{runCohortGeneration}
\title{Run Cohort Generation}
\usage{
runCohortGeneration(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohorts,
  cohortDefinitionSet,
  incrementalFolder = NULL
)
}
\description{
Instantiates all breast cancer cohorts in the target database.
}
