% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{summarizeMeasurements}
\alias{summarizeMeasurements}
\title{Summarize measurements from CohortCounts}
\usage{
summarizeMeasurements(CohortCounts)
}
\description{
Summarize measurements from CohortCounts
}
