% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{summarizeStageDistribution}
\alias{summarizeStageDistribution}
\title{Summarize stage distribution from CohortCounts}
\usage{
summarizeStageDistribution(CohortCounts)
}
\description{
Summarize stage distribution from CohortCounts
}
