% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiagnostics.R
\name{runDiagnostics}
\alias{runDiagnostics}
\title{Run diagnostics for cohorts}
\usage{
runDiagnostics(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  outputFolder = "Results",
  measurement_config = "inst/settings/measurements.json",
  cohortDefinitionSet
)
}
\description{
Run diagnostics for cohorts
}
