% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{summarizeAgeDistribution}
\alias{summarizeAgeDistribution}
\title{Summarize age distribution with customizable bin size}
\usage{
summarizeAgeDistribution(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortDefinitionSet,
  lookup,
  ageBinSize = 10,
  collapseOldestAge = TRUE
)
}
\description{
Summarize age distribution with customizable bin size
}
