% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{summarizeAgeDistribution}
\alias{summarizeAgeDistribution}
\title{Summarize age distribution with 10-year bins}
\usage{
summarizeAgeDistribution(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortDefinitionSet
)
}
\description{
Summarize age distribution with 10-year bins
}
