% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{summarizeCohortCounts}
\alias{summarizeCohortCounts}
\title{Summarize cohort counts}
\usage{
summarizeCohortCounts(
  connectionDetails,
  cohortDatabaseSchema,
  cohortTable,
  cohortDefinitionSet
)
}
\description{
Summarize cohort counts
}
