% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{summarizeMeasurements}
\alias{summarizeMeasurements}
\title{Summarize measurement/receptor positivity (including Unknown) for all cancers}
\usage{
summarizeMeasurements(cohortCounts, measurement_config)
}
\arguments{
\item{cohortCounts}{Data frame returned from summarizeCohortCounts}

\item{measurement_config}{Path to measurement JSON file}
}
\description{
Summarize measurement/receptor positivity (including Unknown) for all cancers
}
