% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{summarizeOverlap}
\alias{summarizeOverlap}
\title{Summarize pairwise overlaps between cohorts}
\usage{
summarizeOverlap(
  connectionDetails,
  cohortDatabaseSchema,
  cohortTable,
  cohortDefinitionSet
)
}
\description{
Summarize pairwise overlaps between cohorts
}
