% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeCohorts.R
\name{summarizeStageDistribution}
\alias{summarizeStageDistribution}
\title{Summarize cancer stage distribution (with missing stages) for all cancers}
\usage{
summarizeStageDistribution(cohortCounts)
}
\description{
Summarize cancer stage distribution (with missing stages) for all cancers
}
